requirements.txt
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-mantine-components==0.12.0
dash-table==5.0.0
dash-tools==1.12.0
dask==2024.10.0
dask-expr==1.1.16
dask-glm==0.3.2
dask-ml==2024.4.4
python-dotenv==1.0.1
gcsfs==2024.10.0
numpy==1.26.4
pandas==2.1.4
plotly==5.15.0
pytz==2024.1
requests==2.31.0
scipy==1.11.4
timezonefinder==6.5.4
statsmodels==0.14.1
scikit-learn==1.4.1.post1
